Move TAS concatenation to a helper function
