+++ dataactbroker/helpers/generic_helper.py
@@ -35,3 +35,25 @@ def quarter_to_dates(quarter):
     end = str(last_month).zfill(2) + '/' + str(last_day_of_month) + '/' + str(year)
 
     return start, end
+
+
+def format_internal_tas(row):
+    """ Concatenate TAS components into a single field for internal use.
+
+        Args:
+            row: row of data with TAS elements
+
+        Returns:
+            TAS components concatenated into a single string
+    """
+    # This formatting should match formatting in dataactcore.models.stagingModels concat_tas
+    tas = ''.join([
+        row['allocation_transfer_agency'] if row['allocation_transfer_agency'] else '000',
+        row['agency_identifier'] if row['agency_identifier'] else '000',
+        row['beginning_period_of_availa'] if row['beginning_period_of_availa'].strip() else '0000',
+        row['ending_period_of_availabil'] if row['ending_period_of_availabil'].strip() else '0000',
+        row['availability_type_code'].strip() if row['availability_type_code'].strip() else ' ',
+        row['main_account_code'] if row['main_account_code'] else '0000',
+        row['sub_account_code'] if row['sub_account_code'] else '000'
+    ])
+    return tas
